+++ app/cdap/services/react/customHooks/useOnUnmount.ts
@@ -1,0 +1,31 @@
+/*
+ * Copyright © 2021 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import { useEffect, useRef } from 'react';
+
+/**
+ * Wrapper for useEffect that ensures the latest version of the callback will be used, unlike useEffect
+ * @param unmountHandler function to call before "unmounting"
+ */
+export function useOnUnmount(unmountHandler: () => any): void {
+  const unmountRef = useRef(unmountHandler);
+  unmountRef.current = unmountHandler;
+  useEffect(() => {
+    return () => {
+      unmountRef.current();
+    };
+  }, [unmountRef]);
+}
