+++ package.json
@@ -69,6 +69,7 @@
     "@storybook/react": "5.2.8",
     "@testing-library/jest-dom": "5.11.5",
     "@testing-library/react": "11.1.0",
+    "@testing-library/react-hooks": "^7.0.2",
     "@types/jest": "26.0.15",
     "@types/react": "16.9.17",
     "@types/react-dom": "16.9.4",
